==> 1_Demographic_mixing_spread_severity/4_Run_final_size_epidemics.R <==
##### RUNS THE UNMITIGATED AND MITIGATED SCENARIOS FOR EACH COUNTRY (NOTE WILL TAKE HOURS SO Global_mitigation_strategies.csv.CSV is already provided)


rm(list=ls())
################# FUNCTIONS TO RUN FINAL SIZES #########################
### functions needed to calculate final sizes###
### outputs the 'error' relative to true final size 
output_set<-function(inf_pops,rates,mixing,demog,inits){
  return(sapply(seq_along(inf_pops),function(x){
    ((demog[x]-inits[x])/inf_pops[x]*(1-exp(-sum(rates[x]*mixing[x,]*(inf_pops-inits)/demog)))-1)
  }))
}


### outputs the sum of squares
sum_output_set<-function(inf_pops,rates,mixing,demog,inits){
  if(min(inf_pops<=0)){return(rep(1000,length(inf_pops)))}
  outset=output_set(inf_pops,rates,mixing,demog,inits)
  return(sum(outset*outset))
}

### homogeneous solution
single_set<-function(inf_pop,rate,pop,init){
  diff=(pop-init)/inf_pop*(1-exp(-rate*(inf_pop-init)/pop))-1
  return(diff*diff)
}


### find the difference the level of herd-immunity generated by a uniform reduction x in contact rates and the threshold level required for Rc<1
diff_from_opt_mit<-function(x,rmod,c_mat,demog,init,guess_hom,guess_modifiers,iterates,reiterates,restarts,tol){
  fit<-get_AR(R=R,rmod=x*rmod,c_mat=c_mat,demog=demog,init=init,guess_hom=guess_hom,guess_modifiers = guess_modifiers,iterates =  iterates,reiterates = reiterates,restarts=restarts,tol=tol)
  si<-fit$par/demog
  reduced_mat<-t(sapply(seq(demog),function(x){
    c_mat[x,]*(1-si)
  }
  ))
  Red_R<-Re(eigen(rmod*reduced_mat)$val[1])
  return((1-Red_R)*(1-Red_R))
}

### same as above but with a fixed reduction "elderly_sd" in 65+
diff_from_opt_mit_elderly_abs<-function(x,elderly_sd,rmod,c_mat,demog,init,guess_hom,guess_modifiers,iterates,reiterates,restarts,tol){
  elderly_dist<-c_mat
  elderly_dist[14:16,]<-elderly_dist[14:16,]*(1-elderly_sd)/x
  fit<-get_AR(R=R,rmod=x*rmod,c_mat=elderly_dist,demog=demog,init=init,guess_hom=guess_hom,guess_modifiers = guess_modifiers,iterates =  iterates,reiterates = reiterates,restarts=restarts,tol=tol)
  si<-fit$par/demog
  reduced_mat<-t(sapply(seq(demog),function(x){
    c_mat[x,]*(1-si)
  }
  ))
  Red_R<-Re(eigen(rmod*reduced_mat)$val[1])
  return((1-Red_R)*(1-Red_R))
}


### get the final size for a set contact matrix and R0
get_AR<-function(R,rmod,c_mat,demog,init,guess_hom,guess_modifiers,iterates,reiterates,restarts,tol){
  ## modifier for next generation matrix to get required R0,1,0.6,5,100000,10,0.00000001
  tot_pop<-sum(demog)
  single_init=init
  ## distribute initial infection across demographics
  init_demog=single_init*demog/tot_pop
  ## get homogenous solution
  fit_err=1
  guess_inf=guess_hom*tot_pop
  ## store final homogenous attack rata
  fit_single<-optim(guess_inf,single_set,rate=R,pop=tot_pop,init=single_init,method="Brent",lower=0,upper=tot_pop,control=list(abstol=tol,maxit=iterates))
  while(fit_err>0.0000001){
    #  print(fit_err)
    ###use homogeouns solution as first  guess  
    fit<-nmkb(fit_single$par*demog/sum(demog)*guess_modifiers, sum_output_set, lower=0, upper=demog,rates=rmod,mixing=c_mat,demog=demog,inits=init_demog,control=list(maxfeval=iterates,tol=tol,restarts.max=restarts))
    for(f in 1:reiterates){
      
      ## reiterate to make sure it finds the solution
      fit<-nmkb(fit$par, sum_output_set, lower=0, upper=demog,rates=rmod,mixing=c_mat,demog=demog,inits=init_demog,control=list(maxfeval=iterates,tol=tol,restarts.max=restarts))
      # print(fit$value)
    }
    fit_err=fit$value
    guess_modifiers=guess_modifiers+0.05
  }
  return(fit)
}


### function to get summary simulations, including mitigation scenarios for given country, R0, contact matrix, demography and IFR, %of hospitals, % critical all by age
get_ARs_IFR<-function(country_name,contact_mat,demog_all,elderly_sd,IFRs,hosp_probs,crit_probs,R,init,guess_hom,guess_modifier,guess_mitigation,reiterates,iterates,restarts,tol){
  
  ### get five year brackets up to 75 and 75+
  demog<-c(demog_all[1:15],sum(demog_all[16:length(demog_all)]))
  
  ## get demand for contacts
  MIJ<-t(sapply(seq(demog),function(x){
    contact_mat[x,]*demog[x]
  }))
  
  ### balance with supply
  adjust_mat<-(MIJ+t(MIJ))/2
  
  ### regain per-capita rates
  new_mix_mat<-t(sapply(seq(demog),function(x){
    adjust_mat[x,]/demog[x]
  }))
  
  ## turn into probability matrix 
  c_mat<-t(sapply(seq(demog),function(x){
    new_mix_mat[x,]/sum(new_mix_mat[x,])
  }
  ))
  ## get overall age-specific rates 
  ai=rowSums(new_mix_mat)
  ### find leading eigen value of next gen mat
  ng_eigen=Re(eigen(new_mix_mat)$val[1])
  rmod<-R/ng_eigen*ai
  ### single initial person for homogenous solution
  single_init=1
  ## total population for homogenous solution
  tot_pop<-sum(demog)
  
  ## distribute initial infection across demographics
  init_demog=single_init*demog/tot_pop
  ## store attributes
  final_vals=array(0,dim=c(length(3)))
  final_err=array(0,dim=c(length(3)))
  age_dists<-array(0,dim=c(3,length(demog)+1))
  age_props<-array(0,dim=c(3,length(demog)))
  deaths_age<-array(0,dim=c(3,length(demog)+1))
  hosps_age<-array(0,dim=c(3,length(demog)+1))
  crits_age<-array(0,dim=c(3,length(demog)+1))
  reductions<-array(0,dim=3)
  IFR<-array(0,dim=c(3))
  hosps<-array(0,dim=c(3))
  crits<-array(0,dim=c(3))
  
  
  
  ### demography with 75-80, 80+ for severity model
  long_demogs<-array(0,dim=c(length(demog)+1))
  long_demogs[1:16]<-demog_all[1:16]
  long_demogs[17]<-sum(demog_all[17:21])
  
  ### find optimal level of mitigation
  fit_mit<-optim(guess_mitigation,diff_from_opt_mit,rmod=rmod,c_mat=c_mat,demog=demog,init=init,guess_hom=guess_hom,guess_modifiers=guess_modifier,iterates=iterates,reiterates=reiterates,restarts=restarts,tol=tol,method="Brent",lower=0.5,upper=0.8)
  fit_mit_elderly_abs<-optim(guess_mitigation,diff_from_opt_mit_elderly_abs,elderly_sd=elderly_sd,rmod=rmod,c_mat=c_mat,demog=demog,init=init,guess_hom=guess_hom,guess_modifiers=guess_modifier,iterates=iterates,reiterates=reiterates,restarts=restarts,tol=tol,method="Brent",lower=0.5,upper=0.8)
  
  ## adjust contact matrix for the mitigation with enhanced SD of elderly
  elderly_dist<-c_mat
  elderly_dist[14:16,]<-elderly_dist[14:16,]*(1-elderly_sd)/fit_mit_elderly_abs$par
  
  
  ### get final sizes for unmitigated, mitigated and mitigated with SD of elderly
  mit_eld<- get_AR(R=R,rmod=fit_mit_elderly_abs$par*rmod,c_mat = elderly_dist,demog = demog,init=init,guess_hom=guess_hom,guess_modifiers = guess_modifier,iterates =  iterates,reiterates = reiterates,restarts=restarts,tol=tol)
  mit<-get_AR(R=R,rmod=fit_mit$par*rmod,c_mat=c_mat,demog = demog,init=init,guess_hom=guess_hom,guess_modifiers = guess_modifier,iterates =  iterates,reiterates = reiterates,restarts=restarts,tol=tol)
  unmit<- get_AR(R=R,rmod=rmod,c_mat=c_mat,demog = demog,init=init,guess_hom=guess_hom,guess_modifiers = guess_modifier,iterates =  iterates,reiterates = reiterates,restarts=restarts,tol=tol)
  
  
  ## store everything
  reductions[2]<-1-fit_mit$par
  reductions[3]<-1-fit_mit_elderly_abs$par
  final_vals[1]<-sum(unmit$par)
  age_dists[1,1:15]<-unmit$par[1:15]
  age_dists[1,16]<-unmit$par[16]*demog_all[16]/sum(demog_all[16:21])
  age_dists[1,17]<-unmit$par[16]*sum(demog_all[17:21])/sum(demog_all[16:21])
  final_err[1]<-unmit$value
  age_props[1,]<-unmit$par/demog
  hosps_age[1,1:15]<-unmit$par[1:15]*hosp_probs[1:15]
  hosps_age[1,16]<-unmit$par[16]*demog_all[16]/sum(demog_all[16:21])*hosp_probs[16]
  hosps_age[1,17]<-unmit$par[16]*sum(demog_all[17:21])/sum(demog_all[16:21])*hosp_probs[17]
  crits_age[1,1:15]<-unmit$par[1:15]*crit_probs[1:15]
  crits_age[1,16]<-unmit$par[16]*demog_all[16]/sum(demog_all[16:21])*crit_probs[16]
  crits_age[1,17]<-unmit$par[16]*sum(demog_all[17:21])/sum(demog_all[16:21])*crit_probs[17]
  deaths_age[1,1:15]<-unmit$par[1:15]*IFRs[1:15]
  deaths_age[1,16]<-unmit$par[16]*demog_all[16]/sum(demog_all[16:21])*IFRs[16]
  deaths_age[1,17]<-unmit$par[16]*sum(demog_all[17:21])/sum(demog_all[16:21])*IFRs[17]
  IFR[1]<-sum(deaths_age[1,])/sum(demog)
  hosps[1]<-sum(hosps_age[1,])/sum(demog)
  crits[1]<-sum(crits_age[1,])/sum(demog)
  reductions[1]<-0
  final_vals[2]<-sum(mit$par)
  age_dists[2,1:15]<-mit$par[1:15]
  age_dists[2,16]<-mit$par[16]*demog_all[16]/sum(demog_all[16:21])
  age_dists[2,17]<-mit$par[16]*sum(demog_all[17:21])/sum(demog_all[16:21])
  final_err[2]<-mit$value
  age_props[2,]<-mit$par/demog
  hosps_age[2,1:15]<-mit$par[1:15]*hosp_probs[1:15]
  hosps_age[2,16]<-mit$par[16]*demog_all[16]/sum(demog_all[16:21])*hosp_probs[16]
  hosps_age[2,17]<-mit$par[16]*sum(demog_all[17:21])/sum(demog_all[16:21])*hosp_probs[17]
  crits_age[2,1:15]<-mit$par[1:15]*crit_probs[1:15]
  crits_age[2,16]<-mit$par[16]*demog_all[16]/sum(demog_all[16:21])*crit_probs[16]
  crits_age[2,17]<-mit$par[16]*sum(demog_all[17:21])/sum(demog_all[16:21])*crit_probs[17]
  deaths_age[2,1:15]<-mit$par[1:15]*IFRs[1:15]
  deaths_age[2,16]<-mit$par[16]*demog_all[16]/sum(demog_all[16:21])*IFRs[16]
  deaths_age[2,17]<-mit$par[16]*sum(demog_all[17:21])/sum(demog_all[16:21])*IFRs[17]
  IFR[2]<-sum(deaths_age[2,])/sum(demog)
  hosps[2]<-sum(hosps_age[2,])/sum(demog)
  crits[2]<-sum(crits_age[2,])/sum(demog)
  final_vals[3]<-sum(mit_eld$par)
  age_dists[3,1:15]<-mit_eld$par[1:15]
  age_dists[3,16]<-mit_eld$par[16]*demog_all[16]/sum(demog_all[16:21])
  age_dists[3,17]<-mit_eld$par[16]*sum(demog_all[17:21])/sum(demog_all[16:21])
  final_err[3]<-mit_eld$value
  age_props[3,]<-mit_eld$par/demog
  hosps_age[3,1:15]<-mit_eld$par[1:15]*hosp_probs[1:15]
  hosps_age[3,16]<-mit_eld$par[16]*demog_all[16]/sum(demog_all[16:21])*hosp_probs[16]
  hosps_age[3,17]<-mit_eld$par[16]*sum(demog_all[17:21])/sum(demog_all[16:21])*hosp_probs[17]
  crits_age[3,1:15]<-mit_eld$par[1:15]*crit_probs[1:15]
  crits_age[3,16]<-mit_eld$par[16]*demog_all[16]/sum(demog_all[16:21])*crit_probs[16]
  crits_age[3,17]<-mit_eld$par[16]*sum(demog_all[17:21])/sum(demog_all[16:21])*crit_probs[17]
  deaths_age[3,1:15]<-mit_eld$par[1:15]*IFRs[1:15]
  deaths_age[3,16]<-mit_eld$par[16]*demog_all[16]/sum(demog_all[16:21])*IFRs[16]
  deaths_age[3,17]<-mit_eld$par[16]*sum(demog_all[17:21])/sum(demog_all[16:21])*IFRs[17]
  IFR[3]<-sum(deaths_age[3,])/sum(demog)
  hosps[3]<-sum(hosps_age[3,])/sum(demog)
  crits[3]<-sum(crits_age[3,])/sum(demog)
  return(list("country"=country_name,"demog"=demog,"R"=R,"rates"=ai/ng_eigen,"reductions"=reductions,"final_size"=final_vals,"age_final_sizes"=age_dists,"age_ARs"=age_props,"deaths_by_age"=deaths_age,"hospitalisations_by_age"=hosps_age,"critical_by_age"=crits_age,"IFR"=IFR,"hospital_prob"=hosps,"critical_prob"=crits,"demog_w_80s"=long_demogs,"errors"=final_err))
}
#################### END OF FINAL SIZE FUNCTIONS ########################################


library(dfoptim)

setwd(dirname(rstudioapi::getActiveDocumentContext()$path))
### get WPP demography
demog_WPP<-read.csv("./Data/country_inputs.csv")
pop_columns<-grep("pop",names(demog_WPP))


### get contact matrices
contact_mat_list<-readRDS("Outputs/contact_matrices.rds")

### get severity inputs
severity_inputs<-read.csv("./Data/severity_inputs.csv",fileEncoding="UTF-8-BOM")


##set R0s
R=2.3
##set a reduction in elderly
elderly_sd=0.6
### initial size
init=1
##pick your countries from the list of available from
countries<-unique(demog_WPP$Country_or_region)

### guess homogenous attack rate
guess_hom=0.85

## guess a lower bound for extent age-specific matrix alters average attack rate
guess_modifier=0.7

### guess level of SD required for optimal mitigation
guess_mitigation=0.65


### optimisation parameters
reiterates=3
iterates=100000
restarts=5
tol=0.0000001
R=2.3
country_sims_2.3<-list()
for(i in seq_along(countries)){
  pick=i
  print(demog_WPP$Country_or_region[pick])
  country_sims_2.3[[i]]<-get_ARs_IFR(country_name = demog_WPP$Country_or_region[pick],contact_mat=data.matrix((contact_mat_list[[demog_WPP$Matrix[pick]]])),demog_all=unlist(demog_WPP[pick,pop_columns]),elderly_sd = elderly_sd,IFRs = severity_inputs$IFR,
                                 hosp_probs =  severity_inputs$hosp_prob,crit_probs = severity_inputs$crit_care,R=R,init=init,
                                 guess_hom =  guess_hom,guess_modifier= guess_modifier, guess_mitigation=guess_mitigation,
                                 reiterates = reiterates,iterates = iterates,restarts = restarts,tol=tol)
}



R=3
country_sims_3.0<-list()
for(i in seq_along(countries)){
  pick=i
  print(demog_WPP$Country_or_region[pick])
  country_sims_3.0[[i]]<-get_ARs_IFR(country_name = demog_WPP$Country_or_region[pick],contact_mat=data.matrix((contact_mat_list[[demog_WPP$Matrix[pick]]])),demog_all=unlist(demog_WPP[pick,pop_columns]),elderly_sd = elderly_sd,IFRs = severity_inputs$IFR,
                                 hosp_probs =  severity_inputs$hosp_prob,crit_probs = severity_inputs$crit_care,R=R,init=init,
                                 guess_hom =  guess_hom,guess_modifier= guess_modifier, guess_mitigation=guess_mitigation,
                                 reiterates = reiterates,iterates = iterates,restarts = restarts,tol=tol)
}



tol=0.00000005
R=3.5 

country_sims_3.5<-list()
for(i in seq_along(countries)){
  pick=i
  print(demog_WPP$Country_or_region[pick])
  country_sims_3.5[[i]]<-get_ARs_IFR(country_name = demog_WPP$Country_or_region[pick],contact_mat=data.matrix((contact_mat_list[[demog_WPP$Matrix[pick]]])),demog_all=unlist(demog_WPP[pick,pop_columns]),elderly_sd = elderly_sd,IFRs = severity_inputs$IFR,
                                     hosp_probs =  severity_inputs$hosp_prob,crit_probs = severity_inputs$crit_care,R=R,init=init,
                                     guess_hom =  guess_hom,guess_modifier= guess_modifier, guess_mitigation=guess_mitigation,
                                     reiterates = reiterates,iterates = iterates,restarts = restarts,tol=tol)
}

seq_along(country_sims_2.3)
country_sims_2.3


### store the output (very messy-sorry!)
lower_ages=(seq(0,80,by=5))
upper_ages=c(seq(5,80,by=5),200)
final_size<-array(dim=c(length(country_sims_2.3),17))
country_demog<-array(dim=c(length(country_sims_2.3),17))
deaths<-array(dim=c(length(country_sims_2.3),17))
hospitalisations<-array(dim=c(length(country_sims_2.3),17))
critical<-array(dim=c(length(country_sims_2.3),17))
country<-array(dim=c(length(country_sims_2.3)))
for(i in seq_along(countries)){
  country_demog[i,]=country_sims_2.3[[i]]$demog_w_80s
  country[i]=country_sims_2.3[[i]]$country
  final_size[i,]=country_sims_2.3[[i]]$age_final_sizes[1,]
  deaths[i,]=country_sims_2.3[[i]]$deaths_by_age[1,]
  long_country_demog[i,]=country_sims_2.3[[i]]$demog_w_80s
  hospitalisations[i,]=country_sims_2.3[[i]]$hospitalisations_by_age[1,]
  critical[i,]=country_sims_2.3[[i]]$critical_by_age[1,]
}

output_df<-data.frame(country,2.3,"Unmitigated",0,rowSums(country_demog),rowSums(final_size),rowSums(deaths),rowSums(hospitalisations),rowSums(critical),country_demog,final_size,deaths,hospitalisations,critical)
names(output_df)<-c("country","R0","Strategy","Reductions","total_pop","total_infected","total_deaths","total_hospital","total_critical",paste("pop",lower_ages,upper_ages,sep="_"),paste("infected",lower_ages,upper_ages,sep="_"),paste("deaths",lower_ages,upper_ages,sep="_"),paste("hospitalisations",lower_ages,upper_ages,sep="_"),paste("critical",lower_ages,upper_ages,sep="_"))


final_size<-array(dim=c(length(country_sims_2.3),17))
country_demog<-array(dim=c(length(country_sims_2.3),17))
deaths<-array(dim=c(length(country_sims_2.3),17))
hospitalisations<-array(dim=c(length(country_sims_2.3),17))
critical<-array(dim=c(length(country_sims_2.3),17))
country<-array(dim=c(length(country_sims_2.3)))
reductions<-array(dim=c(length(country_sims_2.3)))
for(i in seq_along(countries)){
  country_demog[i,]=country_sims_2.3[[i]]$demog_w_80s
  country[i]=country_sims_2.3[[i]]$country
  final_size[i,]=country_sims_2.3[[i]]$age_final_sizes[2,]
  deaths[i,]=country_sims_2.3[[i]]$deaths_by_age[2,]
  long_country_demog[i,]=country_sims_2.3[[i]]$demog_w_80s
  hospitalisations[i,]=country_sims_2.3[[i]]$hospitalisations_by_age[2,]
  critical[i,]=country_sims_2.3[[i]]$critical_by_age[2,]
  reductions[i]=country_sims_2.3[[i]]$reductions[2]
}
output_df_single_R<-data.frame(country,2.3,"Mitigation incl. social distancing",reductions,rowSums(country_demog),rowSums(final_size),rowSums(deaths),rowSums(hospitalisations),rowSums(critical),country_demog,final_size,deaths,hospitalisations,critical)
names(output_df_single_R)<-c("country","R0","Strategy","Reductions","total_pop","total_infected","total_deaths","total_hospital","total_critical",paste("pop",lower_ages,upper_ages,sep="_"),paste("infected",lower_ages,upper_ages,sep="_"),paste("deaths",lower_ages,upper_ages,sep="_"),paste("hospitalisations",lower_ages,upper_ages,sep="_"),paste("critical",lower_ages,upper_ages,sep="_"))
output_df<-rbind(output_df,output_df_single_R)


final_size<-array(dim=c(length(country_sims_2.3),17))
country_demog<-array(dim=c(length(country_sims_2.3),17))
deaths<-array(dim=c(length(country_sims_2.3),17))
hospitalisations<-array(dim=c(length(country_sims_2.3),17))
critical<-array(dim=c(length(country_sims_2.3),17))
country<-array(dim=c(length(country_sims_2.3)))
reductions<-array(dim=c(length(country_sims_2.3)))
for(i in seq_along(countries)){
  country_demog[i,]=country_sims_2.3[[i]]$demog_w_80s
  country[i]=country_sims_2.3[[i]]$country
  final_size[i,]=country_sims_2.3[[i]]$age_final_sizes[3,]
  deaths[i,]=country_sims_2.3[[i]]$deaths_by_age[3,]
  long_country_demog[i,]=country_sims_2.3[[i]]$demog_w_80s
  hospitalisations[i,]=country_sims_2.3[[i]]$hospitalisations_by_age[3,]
  critical[i,]=country_sims_2.3[[i]]$critical_by_age[3,]
  reductions[i]=country_sims_2.3[[i]]$reductions[3]
}
output_df_single_R<-data.frame(country,2.3,"Mitigation incl. enhanced social distancing of elderly",reductions,rowSums(country_demog),rowSums(final_size),rowSums(deaths),rowSums(hospitalisations),rowSums(critical),country_demog,final_size,deaths,hospitalisations,critical)
names(output_df_single_R)<-c("country","R0","Strategy","Reductions","total_pop","total_infected","total_deaths","total_hospital","total_critical",paste("pop",lower_ages,upper_ages,sep="_"),paste("infected",lower_ages,upper_ages,sep="_"),paste("deaths",lower_ages,upper_ages,sep="_"),paste("hospitalisations",lower_ages,upper_ages,sep="_"),paste("critical",lower_ages,upper_ages,sep="_"))
output_df<-rbind(output_df,output_df_single_R)



final_size<-array(dim=c(length(country_sims_3.0),17))
country_demog<-array(dim=c(length(country_sims_3.0),17))
deaths<-array(dim=c(length(country_sims_3.0),17))
hospitalisations<-array(dim=c(length(country_sims_3.0),17))
critical<-array(dim=c(length(country_sims_3.0),17))
country<-array(dim=c(length(country_sims_3.0)))
for(i in seq_along(countries)){
  country_demog[i,]=country_sims_3.0[[i]]$demog_w_80s
  country[i]=country_sims_3.0[[i]]$country
  final_size[i,]=country_sims_3.0[[i]]$age_final_sizes[1,]
  deaths[i,]=country_sims_3.0[[i]]$deaths_by_age[1,]
  hospitalisations[i,]=country_sims_3.0[[i]]$hospitalisations_by_age[1,]
  critical[i,]=country_sims_3.0[[i]]$critical_by_age[1,]
}

output_df_single_R<-data.frame(country,3.0,"Unmitigated",0,rowSums(country_demog),rowSums(final_size),rowSums(deaths),rowSums(hospitalisations),rowSums(critical),country_demog,final_size,deaths,hospitalisations,critical)
names(output_df_single_R)<-c("country","R0","Strategy","Reductions","total_pop","total_infected","total_deaths","total_hospital","total_critical",paste("pop",lower_ages,upper_ages,sep="_"),paste("infected",lower_ages,upper_ages,sep="_"),paste("deaths",lower_ages,upper_ages,sep="_"),paste("hospitalisations",lower_ages,upper_ages,sep="_"),paste("critical",lower_ages,upper_ages,sep="_"))
output_df<-rbind(output_df,output_df_single_R)



final_size<-array(dim=c(length(country_sims_3.0),17))
country_demog<-array(dim=c(length(country_sims_3.0),17))
deaths<-array(dim=c(length(country_sims_3.0),17))
hospitalisations<-array(dim=c(length(country_sims_3.0),17))
critical<-array(dim=c(length(country_sims_3.0),17))
country<-array(dim=c(length(country_sims_3.0)))
reductions<-array(dim=c(length(country_sims_3.0)))
for(i in seq_along(countries)){
  country_demog[i,]=country_sims_3.0[[i]]$demog_w_80s
  country[i]=country_sims_3.0[[i]]$country
  final_size[i,]=country_sims_3.0[[i]]$age_final_sizes[2,]
  deaths[i,]=country_sims_3.0[[i]]$deaths_by_age[2,]
  long_country_demog[i,]=country_sims_3.0[[i]]$demog_w_80s
  hospitalisations[i,]=country_sims_3.0[[i]]$hospitalisations_by_age[2,]
  critical[i,]=country_sims_3.0[[i]]$critical_by_age[2,]
  reductions[i]=country_sims_3.0[[i]]$reductions[2]
}
output_df_single_R<-data.frame(country,3.0,"Mitigation incl. social distancing",reductions,rowSums(country_demog),rowSums(final_size),rowSums(deaths),rowSums(hospitalisations),rowSums(critical),country_demog,final_size,deaths,hospitalisations,critical)
names(output_df_single_R)<-c("country","R0","Strategy","Reductions","total_pop","total_infected","total_deaths","total_hospital","total_critical",paste("pop",lower_ages,upper_ages,sep="_"),paste("infected",lower_ages,upper_ages,sep="_"),paste("deaths",lower_ages,upper_ages,sep="_"),paste("hospitalisations",lower_ages,upper_ages,sep="_"),paste("critical",lower_ages,upper_ages,sep="_"))
output_df<-rbind(output_df,output_df_single_R)


final_size<-array(dim=c(length(country_sims_3.0),17))
country_demog<-array(dim=c(length(country_sims_3.0),17))
deaths<-array(dim=c(length(country_sims_3.0),17))
hospitalisations<-array(dim=c(length(country_sims_3.0),17))
critical<-array(dim=c(length(country_sims_3.0),17))
country<-array(dim=c(length(country_sims_3.0)))
reductions<-array(dim=c(length(country_sims_3.0)))
for(i in seq_along(countries)){
  country_demog[i,]=country_sims_3.0[[i]]$demog_w_80s
  country[i]=country_sims_3.0[[i]]$country
  final_size[i,]=country_sims_3.0[[i]]$age_final_sizes[3,]
  deaths[i,]=country_sims_3.0[[i]]$deaths_by_age[3,]
  long_country_demog[i,]=country_sims_3.0[[i]]$demog_w_80s
  hospitalisations[i,]=country_sims_3.0[[i]]$hospitalisations_by_age[3,]
  critical[i,]=country_sims_3.0[[i]]$critical_by_age[3,]
  reductions[i]=country_sims_3.0[[i]]$reductions[3]
}
output_df_single_R<-data.frame(country,3.0,"Mitigation incl. enhanced social distancing of elderly",reductions,rowSums(country_demog),rowSums(final_size),rowSums(deaths),rowSums(hospitalisations),rowSums(critical),country_demog,final_size,deaths,hospitalisations,critical)
names(output_df_single_R)<-c("country","R0","Strategy","Reductions","total_pop","total_infected","total_deaths","total_hospital","total_critical",paste("pop",lower_ages,upper_ages,sep="_"),paste("infected",lower_ages,upper_ages,sep="_"),paste("deaths",lower_ages,upper_ages,sep="_"),paste("hospitalisations",lower_ages,upper_ages,sep="_"),paste("critical",lower_ages,upper_ages,sep="_"))
output_df<-rbind(output_df,output_df_single_R)




final_size<-array(dim=c(length(country_sims_3.5),17))
country_demog<-array(dim=c(length(country_sims_3.5),17))
deaths<-array(dim=c(length(country_sims_3.5),17))
hospitalisations<-array(dim=c(length(country_sims_3.5),17))
critical<-array(dim=c(length(country_sims_3.5),17))
country<-array(dim=c(length(country_sims_3.5)))
for(i in seq_along(countries)){
  country_demog[i,]=country_sims_3.5[[i]]$demog_w_80s
  country[i]=country_sims_3.5[[i]]$country
  final_size[i,]=country_sims_3.5[[i]]$age_final_sizes[1,]
  deaths[i,]=country_sims_3.5[[i]]$deaths_by_age[1,]
  hospitalisations[i,]=country_sims_3.5[[i]]$hospitalisations_by_age[1,]
  critical[i,]=country_sims_3.5[[i]]$critical_by_age[1,]
}

output_df_single_R<-data.frame(country,3.5,"Unmitigated",0,rowSums(country_demog),rowSums(final_size),rowSums(deaths),rowSums(hospitalisations),rowSums(critical),country_demog,final_size,deaths,hospitalisations,critical)
names(output_df_single_R)<-c("country","R0","Strategy","Reductions","total_pop","total_infected","total_deaths","total_hospital","total_critical",paste("pop",lower_ages,upper_ages,sep="_"),paste("infected",lower_ages,upper_ages,sep="_"),paste("deaths",lower_ages,upper_ages,sep="_"),paste("hospitalisations",lower_ages,upper_ages,sep="_"),paste("critical",lower_ages,upper_ages,sep="_"))
output_df<-rbind(output_df,output_df_single_R)



final_size<-array(dim=c(length(country_sims_3.5),17))
country_demog<-array(dim=c(length(country_sims_3.5),17))
deaths<-array(dim=c(length(country_sims_3.5),17))
hospitalisations<-array(dim=c(length(country_sims_3.5),17))
critical<-array(dim=c(length(country_sims_3.5),17))
country<-array(dim=c(length(country_sims_3.5)))
reductions<-array(dim=c(length(country_sims_3.5)))
for(i in seq_along(countries)){
  country_demog[i,]=country_sims_3.5[[i]]$demog_w_80s
  country[i]=country_sims_3.5[[i]]$country
  final_size[i,]=country_sims_3.5[[i]]$age_final_sizes[2,]
  deaths[i,]=country_sims_3.5[[i]]$deaths_by_age[2,]
  hospitalisations[i,]=country_sims_3.5[[i]]$hospitalisations_by_age[2,]
  critical[i,]=country_sims_3.5[[i]]$critical_by_age[2,]
  reductions[i]=country_sims_3.5[[i]]$reductions[2]
}
output_df_single_R<-data.frame(country,3.5,"Mitigation incl. social distancing",reductions,rowSums(country_demog),rowSums(final_size),rowSums(deaths),rowSums(hospitalisations),rowSums(critical),country_demog,final_size,deaths,hospitalisations,critical)
names(output_df_single_R)<-c("country","R0","Strategy","Reductions","total_pop","total_infected","total_deaths","total_hospital","total_critical",paste("pop",lower_ages,upper_ages,sep="_"),paste("infected",lower_ages,upper_ages,sep="_"),paste("deaths",lower_ages,upper_ages,sep="_"),paste("hospitalisations",lower_ages,upper_ages,sep="_"),paste("critical",lower_ages,upper_ages,sep="_"))
output_df<-rbind(output_df,output_df_single_R)


final_size<-array(dim=c(length(country_sims_3.5),17))
country_demog<-array(dim=c(length(country_sims_3.5),17))
deaths<-array(dim=c(length(country_sims_3.5),17))
hospitalisations<-array(dim=c(length(country_sims_3.5),17))
critical<-array(dim=c(length(country_sims_3.5),17))
country<-array(dim=c(length(country_sims_3.5)))
reductions<-array(dim=c(length(country_sims_3.5)))
for(i in seq_along(countries)){
  country_demog[i,]=country_sims_3.5[[i]]$demog_w_80s
  country[i]=country_sims_3.5[[i]]$country
  final_size[i,]=country_sims_3.5[[i]]$age_final_sizes[3,]
  deaths[i,]=country_sims_3.5[[i]]$deaths_by_age[3,]
  long_country_demog[i,]=country_sims_3.5[[i]]$demog_w_80s
  hospitalisations[i,]=country_sims_3.5[[i]]$hospitalisations_by_age[3,]
  critical[i,]=country_sims_3.5[[i]]$critical_by_age[3,]
  reductions[i]=country_sims_3.5[[i]]$reductions[3]
}
output_df_single_R<-data.frame(country,3.5,"Mitigation incl. enhanced social distancing of elderly",reductions,rowSums(country_demog),rowSums(final_size),rowSums(deaths),rowSums(hospitalisations),rowSums(critical),country_demog,final_size,deaths,hospitalisations,critical)
names(output_df_single_R)<-c("country","R0","Strategy","Reductions","total_pop","total_infected","total_deaths","total_hospital","total_critical",paste("pop",lower_ages,upper_ages,sep="_"),paste("infected",lower_ages,upper_ages,sep="_"),paste("deaths",lower_ages,upper_ages,sep="_"),paste("hospitalisations",lower_ages,upper_ages,sep="_"),paste("critical",lower_ages,upper_ages,sep="_"))
output_df<-rbind(output_df,output_df_single_R)

write.csv(output_df,"Outputs/Global_mitigation_strategies_redo.csv",row.names=F)
